refactor and rename runner.py to be test environment specific
Also create and move temporary_env to utils.py
